multichip support, approach 2 discussed in PR
