--- rnfc-st25r39/src/lib_st25r3916.rs
@@ -1,6 +1,0 @@
-// TODO - placeholder if "approach 2" described in PR is preferred,
-// then specific for st25r3916 parts from "lib" could be moved here
-// which would allow to align the changes between chips better,
-// and possibly remove substantial amount of "cfg!" checks
-// See also discussion on Matrix
-// https://matrix.to/#/!YoLPkieCYHGzdjUhOK:matrix.org/$jLJeWLyptxa4kbQcl-j2hKoha9DeI_Ta6m_X0t2wmhw
